+++ pkgs/unified_analytics/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 5.8.1
 
 - Refactor logic for `okToSend` and `shouldShowMessage`
+- Check devtools config file for legacy opt out status
 
 ## 5.8.0